Pass configs in localrc to the ovs-install script.
This commit enhances the quick installer. It is a preparation for
multi node support with the installer. Now scripts/install-ovs-as-ofs.sh
reads some configuration such as OFC_HOST and OVS_INTERFACE from localrc
used by devstack. OFC_HOST specifies a OpenFlow controler IP address.
If OVS_INTERFACE is defined as a non-empty value, an interface specified
by this variable is attached to OVS_BRIDGE (br-int) in the script.